calculo da necessidade em peca baseado na materia prima, corrigindo para carregar csv

+++ src/models/Tendencia_Plano_Materiais.py
@@ -379,7 +379,7 @@ class Tendencia_Plano_Materiais():
         sqlEstoque = sqlEstoque.groupby(["CodComponente"]).agg(
             {"estoqueAtual": "sum"}).reset_index()
 
-        Necessidade['CodComponente'] = Necessidade['CodComponente'].astype(str)
+        Necessidade['CodComponente'] = Necessidade['CodComponente'].astype(float).astype(int).astype(str)
         Necessidade = pd.merge(Necessidade, sqlEstoque, on='CodComponente', how='left')
 
         Necessidade.to_csv(f'{caminho_absoluto2}/dados/MeuTeste2.csv')
